requirements.txt
aiofiles==0.4.0
alabaster==0.7.8
ansible==2.9.4
ansible-runner==1.4.4
appdirs==1.4.3
argcomplete==1.8.1
asn1crypto==0.24.0
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.6.0
black==19.3b0
blessings==1.6
blinker==1.4
bpython==0.18
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
curtsies==0.3.1
Cython==0.29.14
dbus-python==1.2.16
dbus2vdr==0.0.1
DBussy==1.2.1
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
fastapi==0.52.0
flake8==3.7.9
greenlet==0.4.15
h11==0.9.0
httplib2==0.14.0
httptools==0.1.1
hyperlink==19.0.0
idna==2.6
imagesize==1.2.0
incremental==16.10.1
Jinja2==2.10.1
jmespath==0.9.4
keyring==18.0.1
keyrings.alt==3.2.0
kmodpy==0.1.13
language-selector==0.1
launchpadlib==1.10.10
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
lxml==4.5.0
lz4==3.0.2+dfsg
MarkupSafe==1.1.0
mccabe==0.6.1
mysql-connector-python==8.0.15
netaddr==0.7.19
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
packaging==20.1
pexpect==4.8.0
pkg-resources==0.0.0
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.6.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycrypto==2.6.1
pydantic==1.4
pydbus==0.6.0
pydbus2vdr==0.0.8
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.3.1
PyGObject==3.34.0
PyHamcrest==1.9.0
pyhton3-avahi==0.0.1
PyJWT==1.7.1
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQRCode==1.2.1
python-apt==1.9.7
python-daemon==2.2.4
python-debian===0.1.36build1
python-multipart==0.0.5
python-snappy==0.5.3
python-uinput==0.11.2
PyTrie==0.2
pytz==2019.3
pyusb==1.0.2
pyxdg==0.26
PyYAML==5.3
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
scour==0.37
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
Sphinx==1.8.5
ssh-import-id==5.10
starlette==0.13.2
systemd-python==234
toml==0.10.0
Twisted==18.9.0
txaio==2.10.0
u-msgpack-python==2.1
ubuntu-advantage-tools==20.2
ubuntu-drivers-common==0.0.0
udiskie==2.1.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
wadllib==1.3.3
wcwidth==0.1.8
websockets==8.1
wsaccel==0.6.2
xkit==0.0.0
yavdr-frontend==0.0.63
zope.interface==4.7.1
